refactor(example-app): split changes and change lists state into two files
