+++ packages/example-app/src/VersionControl/Changes/changes.state.ts
@@ -1,0 +1,146 @@
+import { selector, selectorFamily, useRecoilCallback } from "recoil";
+import path from "path";
+import { groupBy } from "ramda";
+import { checkout, resetIndex } from "isomorphic-git";
+import { notNull } from "@intellij-platform/core/utils/array-utils";
+
+import { dirContentState, reloadFileFromDiskCallback } from "../../fs/fs.state";
+import { fs } from "../../fs/fs";
+import {
+  repoStatusState,
+  useRefreshFileStatus,
+  vcsRootForFile,
+  vcsRootsState,
+} from "../file-status.state";
+import { findRootPaths } from "../../path-utils";
+import { Change } from "./Change";
+
+const repoChangesState = selectorFamily<Change[], string>({
+  key: "vcs.repoChanges",
+  get:
+    (repoDir: string) =>
+    ({ get }) =>
+      Object.entries(get(repoStatusState(repoDir)))
+        .map(([filename, status]) => {
+          const revision = {
+            path: path.join(repoDir, filename),
+            isDir: false,
+            content(): Promise<string> {
+              throw new Error("Not implemented");
+            },
+          };
+          if (status === "ADDED") {
+            return {
+              after: revision,
+            };
+          }
+          if (status === "MODIFIED") {
+            return {
+              before: revision,
+              after: revision,
+            };
+          }
+          if (status === "DELETED") {
+            return {
+              before: revision,
+            };
+          }
+          return null;
+        })
+        .filter(notNull),
+});
+
+export const allChangesState = selector<ReadonlyArray<Change>>({
+  key: "vcs.allChanges",
+  get: ({ get }) => {
+    return get(vcsRootsState)
+      .map(({ dir }) => get(repoChangesState(dir)))
+      .flat();
+  },
+});
+
+/**
+ * react hook that returns a rollback function which accepts a list of changes to rollback.
+ */
+export const useRollbackChanges = () => {
+  const updateFileStatus = useRefreshFileStatus();
+  return useRecoilCallback(
+    (callbackInterface) =>
+      async (changes: readonly Change[], deleteAddedFiles?: boolean) => {
+        const { snapshot, refresh } = callbackInterface;
+        const reloadFileFromDisk =
+          reloadFileFromDiskCallback(callbackInterface);
+
+        const changesWithRepoRoots = (
+          await Promise.all(
+            changes
+              .filter((change) => !change.after?.isDir)
+              .map(async (change) => {
+                const repoRoot = await snapshot.getPromise(
+                  vcsRootForFile(Change.path(change))
+                );
+                if (repoRoot) {
+                  return {
+                    repoRoot,
+                    type: Change.type(change),
+                    fullPath: Change.path(change),
+                    relativePath: path.relative(repoRoot, Change.path(change)),
+                  };
+                }
+              })
+          )
+        ).filter(notNull);
+        const groupedChanges = groupBy(
+          ({ repoRoot }) => repoRoot,
+          changesWithRepoRoots
+        );
+        await Promise.all(
+          Object.entries(groupedChanges).map(async ([repoRoot, items]) => {
+            const { toReset = [], toCheckout = [] } = groupBy(
+              ({ type }) =>
+                type !== "ADDED" || deleteAddedFiles ? "toCheckout" : "toReset",
+              items
+            );
+            await Promise.allSettled(
+              toReset.map(({ relativePath, type }) =>
+                resetIndex({
+                  fs,
+                  dir: repoRoot,
+                  filepath: relativePath,
+                })
+              )
+            );
+            if (toCheckout.length > 0) {
+              await checkout({
+                fs,
+                dir: repoRoot,
+                force: true,
+                filepaths: toCheckout.map(
+                  ({ relativePath, type }) => relativePath
+                ),
+              });
+            }
+
+            // FIXME(fs.watch)
+            const dirsWithRemovedFiles = findRootPaths(
+              items
+                .filter(({ type }) => deleteAddedFiles && type === "ADDED")
+                .map(({ fullPath }) => path.dirname(fullPath))
+            );
+            dirsWithRemovedFiles.forEach((pathToRefresh) =>
+              refresh(dirContentState(pathToRefresh))
+            );
+
+            await Promise.allSettled(
+              items.map(async ({ fullPath, relativePath }) => {
+                await reloadFileFromDisk(fullPath); // Since fileContent is an atom, we set the value. Could be a selector that we would refresh
+                await updateFileStatus(fullPath);
+                await resetIndex({ fs, dir: repoRoot, filepath: relativePath });
+              })
+            );
+          })
+        );
+      },
+    []
+  );
+};
